Add cmake project files, add submodules, add gitignore, meshes and fix some compile errors and warnings

+++ Application/Rendering/MemoryUtils.h
@@ -24,7 +24,7 @@ namespace memory
 	template<typename type>
 	void deallocArray( type* ptr, size_t size )
 	{
-		for (int i = 0; i < size; i++)
+		for (size_t i = 0; i < size; i++)
 		{
 			ptr[i].~type();
 		}
